+++ tools/update_about.py
@@ -1,24 +1,9 @@
 """Generate project metadata and dump it to `about.py`."""  # noqa: INP001
+from pathlib import Path
 import sys
 from textwrap import dedent
-from typing import TYPE_CHECKING
 
-from legion import ensure_utf8_output, excepthook, generate_metadata_file, resolve_version, timestamp
-
-if TYPE_CHECKING:
-    from pathlib import Path
-    from typing import Any
-
-
-TEMPLATE = dedent(f'''
-    """Auto generated, do not edit."""
-    # {timestamp('%Y-%m-%d %H:%M:%S')}
-
-    # pylint: disable=unused-variable
-    PROGRAM_NAME = {{project[name]!r}}
-    VERSION = {{project[version]!r}}
-    RELEASE = {{release!r}}
-''').lstrip()
+from legion import ensure_utf8_output, excepthook, get_project_metadata, resolve_version
 
 
 @ensure_utf8_output
@@ -27,18 +12,23 @@ def main() -> int | str:
     # ruff: disable[T201]
     sys.excepthook = excepthook
 
-    def output_path_factory(m: dict[str, Any]) -> Path:
-        return m['project_root'] / 'src' / m['project']['name'] / 'about.py'
-
     if (release := resolve_version('{tag}.{distance}')) is None:
         return 'Release cannot be determined.'
     tag, _, distance = release.rpartition('.')
 
-    extra_metadata = {'release': tag if distance == '0' else f'{tag}.post0'}
-    if (about_py_file := generate_metadata_file(output_path_factory, TEMPLATE, extra_metadata)) is None:
-        return f"Error generating metadata file at '{about_py_file}'."
+    extra_metadata = {
+        'release': tag if distance == '0' else f'{tag}.post0',
+    }
+    if (project_metadata := get_project_metadata()) is None:
+        return 'Error geting project metadata.'
 
-    print(f"Generated metadata file at '{about_py_file}'")
+    output_path = Path(project_metadata['project_root']) / 'src' / project_metadata['project']['name'] / 'about.py'
+    template = dedent(project_metadata['tool']['metadata']['template'])
+
+    extra_metadata = {'release': tag if distance == '0' else f'{tag}.post0'}
+    output_path.resolve().write_text(template.format_map(project_metadata | extra_metadata))
+
+    print(f"Generated metadata file at '{output_path}'")
     return 0
     # ruff: enable[T201]